[FIX] 스크롤 이벤트 주간 달력 없어지지 않도록 수정
