+++ fairer-iOS/fairer-iOS/Screens/Home/UIComponent/DailyCollectionView/CalendarDailyTableViewCell.swift
@@ -92,14 +92,13 @@ final class CalendarDailyTableViewCell: BaseTableViewCell {
     }()
     
     // MARK: - life cycle
-    
     override func layoutSubviews() {
         super.layoutSubviews()
-
         contentView.frame = contentView.frame.inset(by: UIEdgeInsets(top: 0, left: 0, bottom: 16, right: 0))
     }
     
     override func render(){
+        
         self.addSubviews(leftStackView,rightStackView)
         
         workerImage1.snp.makeConstraints {
@@ -119,12 +118,12 @@ final class CalendarDailyTableViewCell: BaseTableViewCell {
         }
         
         leftStackView.snp.makeConstraints {
-            $0.top.equalToSuperview().offset(SizeLiteral.componentPadding)
-            $0.leading.equalToSuperview().offset(SizeLiteral.componentPadding)
+            $0.top.equalToSuperview().inset(SizeLiteral.componentPadding)
+            $0.leading.equalToSuperview().inset(SizeLiteral.componentPadding)
         }
 
         rightStackView.snp.makeConstraints {
-            $0.top.equalToSuperview().offset(SizeLiteral.componentPadding)
+            $0.top.equalToSuperview().inset(SizeLiteral.componentPadding)
             $0.trailing.equalToSuperview().inset(SizeLiteral.componentPadding)
         }
     }
