+++ fairer-iOS/fairer-iOS/Screens/HomeViewController.swift
@@ -9,10 +9,18 @@ import UIKit
 
 import SnapKit
 
+struct dummyWorkCard {
+    let work = ["바닥 청소","설거지","빨래","바닥 청소","설거지","빨래","설거지"]
+    let time = ["오전 9:30","오후 8:00","오전 11:00","오전 9:30","오후 8:00","오전 11:00","오전 11:00"]
+    let room = ["방","부엌","거실","방","부엌","방","부엌"]
+    let status = [ WorkState.overdue,WorkState.overdue,WorkState.notFinished,WorkState.finished,WorkState.finished,WorkState.finished,WorkState.finished]
+}
+
 final class HomeViewController: BaseViewController {
     
     // TODO: - 추후 api연결 + UserDefault
     
+    private let dummy = dummyWorkCard()
     let userName: String = "고가혜"
     let ruleArray: [String] = ["설거지는 바로바로", "신발 정리하기", "화분 물주기", "밥 다먹은 사람이 치우기"]
     private var isScrolled = false
@@ -20,8 +28,7 @@ final class HomeViewController: BaseViewController {
     private lazy var rightSwipeGestureRecognizer = UISwipeGestureRecognizer(target: self, action: #selector(handleSwipes(_:)))
     
     // MARK: - property
-    
-    private var cellHeight = CGFloat()
+
     private let logoImage = UIImageView(image: ImageLiterals.imgHomeLogo)
     private let profileButton: UIButton = {
         let button = UIButton(type: .system)
@@ -61,7 +68,15 @@ final class HomeViewController: BaseViewController {
     }()
     private let homeCalenderView = HomeCalendarView()
     private let homeWeekCalendarCollectionView = HomeWeekCalendarCollectionView()
-    private let calendarDailyTableView = CalendarDailyTableView()
+    private let calendarDailyTableView: UITableView = {
+        let calendarDailyTableView = UITableView()
+        calendarDailyTableView.register(CalendarDailyTableViewCell.self, forCellReuseIdentifier: CalendarDailyTableViewCell.identifier)
+        calendarDailyTableView.showsVerticalScrollIndicator = false
+        calendarDailyTableView.separatorStyle = .none
+        calendarDailyTableView.rowHeight = 94
+        calendarDailyTableView.backgroundColor = .white
+        return calendarDailyTableView
+    }()
     private lazy var contentScrollView: UIScrollView = {
         let scrollView = UIScrollView()
         scrollView.backgroundColor = .systemBackground
@@ -89,7 +104,6 @@ final class HomeViewController: BaseViewController {
 
     override func configUI() {
         super.configUI()
-        setCalendarHeight()
         setupToolBarGesture()
         setHomeRuleLabel()
     }
@@ -223,6 +237,8 @@ final class HomeViewController: BaseViewController {
     
     private func setupDelegate() {
         self.contentScrollView.delegate = self
+        self.calendarDailyTableView.delegate = self
+        self.calendarDailyTableView.dataSource = self
     }
     
     private func setupToolBarGesture() {
@@ -347,19 +363,6 @@ final class HomeViewController: BaseViewController {
         self.datePickerView.isHidden = false
         self.setupAlphaNavigationBar()
     }
-    
-    private func setCalendarHeight() {
-        self.calendarDailyTableView.changeHeightClosure = { [weak self]
-            cellNum in
-            guard let self = self else {
-                return
-            }
-            self.cellHeight = CGFloat(cellNum) * SizeLiteral.homeViewWorkCellHeight
-            self.calendarDailyTableView.snp.updateConstraints {
-                $0.height.equalTo(self.cellHeight)
-            }
-        }
-    }
 }
 
     // MARK: - extension
@@ -378,3 +381,55 @@ extension HomeViewController: UIScrollViewDelegate {
         }
     }
 }
+
+extension HomeViewController: UITableViewDelegate {}
+extension HomeViewController: UITableViewDataSource {
+    
+    func numberOfSections(in tableView: UITableView) -> Int {
+        return 2
+    }
+
+    func tableView(_ tableView: UITableView, titleForHeaderInSection section: Int) -> String? {
+        if section == 0 {
+            return ""
+        }else {
+            return "끝낸 집안일"
+        }
+    }
+    
+    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+            var cellNum = Int()
+            if section == 0 {
+                cellNum = cellNum + 3
+                return 3
+            }else {
+            cellNum = cellNum + 4
+            let cellHeight = CGFloat(cellNum) * SizeLiteral.homeViewWorkCellHeight
+            self.calendarDailyTableView.snp.updateConstraints {
+                $0.height.equalTo(cellHeight)
+            }
+            cellNum = 0
+            return 4
+        }
+    }
+    
+    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        let cell = calendarDailyTableView.dequeueReusableCell(withIdentifier: CalendarDailyTableViewCell.identifier, for: indexPath) as? CalendarDailyTableViewCell ?? CalendarDailyTableViewCell()
+        cell.selectionStyle = .none
+        cell.workLabel.text = dummy.work[indexPath.item]
+        cell.time.text = dummy.time[indexPath.item]
+        cell.room.text = dummy.room[indexPath.item]
+        switch dummy.status[indexPath.item] {
+        case .finished :
+            cell.backgroundColor = .positive10
+        case .notFinished :
+            cell.backgroundColor = .white
+        case .overdue :
+            cell.backgroundColor = .negative0
+            cell.layer.borderColor = UIColor.negative10.cgColor
+            cell.setErrorImageView()
+        }
+        return cell
+    }
+}
+
